add admin acces for delet methods

+++ views/items.py
@@ -1,6 +1,6 @@
-import sqlite3
 from flask_restful import Resource,reqparse
-from flask_jwt import JWT , jwt_required
+from flask_jwt_extended import jwt_required,create_access_token
+
 from models.items import ItemModel
 class Item(Resource):
     def __init__(self):
@@ -16,7 +16,7 @@ class Item(Resource):
         help="every item need store id "
         )
     
-
+    @jwt_required
     def get(self,name):
         item =  ItemModel.find_by_name(name)
         if item : 
@@ -36,10 +36,13 @@ class Item(Resource):
         except:
             return {"msg":"An error occurred inserting the item"}
        
-
+    @jwt_required
     def delete (self,name):
+        additional_claims = {"id": 1}
+        access_token = create_access_token(name, additional_claims=additional_claims)
         item = ItemModel.find_by_name(name)
-        
+        if not access_token['is_admin']:
+            return{"msg": "your don't have permission "},401
         if item:
             item.delete_from_db()
             return {"msg":"item was deleted"}
